# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AI FE Exam API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/companies": {
            "get": {
                "tags": [
                    "Companies"
                ],
                "summary": "Get all companies",
                "operationId": "80eb39cf3f8d7b480171fea2f33d6060",
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Company"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Companies retrieved successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Companies"
                ],
                "summary": "Create a new company",
                "operationId": "cee123b1da67beec88470f8ceb3fcb95",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "logo",
                                    "name",
                                    "status"
                                ],
                                "properties": {
                                    "logo": {
                                        "description": "Logo of the company",
                                        "type": "string",
                                        "format": "url",
                                        "example": "https://example.com/logo.jpg"
                                    },
                                    "name": {
                                        "description": "Name of the company",
                                        "type": "string",
                                        "example": "Company"
                                    },
                                    "status": {
                                        "description": "Status of the company",
                                        "type": "string",
                                        "example": "Active"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Company"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Company created successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation Error."
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": "Validation Error."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/companies/{id}": {
            "get": {
                "tags": [
                    "Companies"
                ],
                "summary": "Get a company",
                "operationId": "fd3858af6b6071f4a94538a412bab106",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the company",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Company"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Company retrieved successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Companies"
                ],
                "summary": "Update a company",
                "operationId": "c125e6ebfb9c7669666225b619092333",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the company",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "logo",
                                    "name",
                                    "status"
                                ],
                                "properties": {
                                    "logo": {
                                        "description": "Logo of the company",
                                        "type": "string",
                                        "format": "url",
                                        "example": "https://example.com/logo.jpg"
                                    },
                                    "name": {
                                        "description": "Name of the company",
                                        "type": "string",
                                        "example": "Company"
                                    },
                                    "status": {
                                        "description": "Status of the company",
                                        "type": "string",
                                        "example": "Active"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Company"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Company updated successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation Error."
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": "Validation Error."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Companies"
                ],
                "summary": "Delete a company",
                "operationId": "124aa61ff352c83ede5872644b0ac00d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the company",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Company deleted successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Company": {
                "title": "Company",
                "description": "Company model",
                "required": [
                    "logo",
                    "name",
                    "status"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "logo": {
                        "type": "string",
                        "example": "logo"
                    },
                    "name": {
                        "type": "string",
                        "example": "name"
                    },
                    "status": {
                        "type": "string",
                        "example": "status"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2021-09-01T00:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2021-09-01T00:00:00.000000Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}